requirements.txt
streamlit==1.37.0
pandas==2.2.2
pyyaml==6.0.2
plotly==5.22.0
numpy==1.26.4
